AstroQueryGPT/requirements.txt
openai
streamlit
pandas
requests
plotly
python-dotenv
sentence-transformers==2.2.2
